Fix bad gradient checks
